Extract offline sync code from service
