+++ stream-chat-android-offline/api/stream-chat-android-offline.api
@@ -300,6 +300,12 @@ public final class io/getstream/chat/android/livedata/repository/domain/user/Use
 	public fun selectUsersLikeName (Ljava/lang/String;II)Ljava/util/List;
 }
 
+public final class io/getstream/chat/android/livedata/service/sync/StreamOfflineSyncFirebaseServiceDelegate {
+	public fun <init> (Landroid/app/Service;)V
+	public final fun onMessageReceived (Lcom/google/firebase/messaging/RemoteMessage;)V
+	public final fun onNewToken (Ljava/lang/String;)V
+}
+
 public abstract interface class io/getstream/chat/android/livedata/usecase/CancelMessage {
 	public abstract fun invoke (Lio/getstream/chat/android/client/models/Message;)Lio/getstream/chat/android/client/call/Call;
 }
